bpf: Keep cgroup attach storage ABI stable
BPF_CGROUP_INET_SOCK_RELEASE uses sparse UAPI value 34. Sizing the
cgroup_bpf arrays from MAX_BPF_ATTACH_TYPE grows struct cgroup and changes
the module_layout modversion CRC.

Map UAPI attach types to a dense cgroup-private enum while retaining the
original 23-entry array dimensions. Keep the existing bpf_attach_type
enum and legacy sentinel unchanged for modversion compatibility, and expose
the new hook as a typed constant with a separate full attach-type bound.

Teach libbpf and bpftool to load, attach, and report cgroup/sock_release.

+++ kernel/bpf/syscall.c
@@ -1705,7 +1705,7 @@ bpf_prog_load_check_attach(enum bpf_prog_type prog_type,
 {
 	switch (prog_type) {
 	case BPF_PROG_TYPE_CGROUP_SOCK:
-		switch (expected_attach_type) {
+		switch ((int)expected_attach_type) {
 		case BPF_CGROUP_INET_SOCK_CREATE:
 		case BPF_CGROUP_INET_SOCK_RELEASE:
 		case BPF_CGROUP_INET4_POST_BIND:
