+++ net/core/filter.c
@@ -6732,7 +6732,7 @@ static bool __sock_filter_check_attach_type(int off,
 	case offsetof(struct bpf_sock, bound_dev_if):
 	case offsetof(struct bpf_sock, mark):
 	case offsetof(struct bpf_sock, priority):
-		switch (attach_type) {
+		switch ((int)attach_type) {
 		case BPF_CGROUP_INET_SOCK_CREATE:
 		case BPF_CGROUP_INET_SOCK_RELEASE:
 			goto full_access;
